I have extracted DTOs to the separate library.
