+++ ETIM-API-Saver/Common/DTO/Include.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Common.DTO
+{
+    public class Include
+    {
+        // true
+        public bool Descriptions { get; set; }
+        // false
+        public bool Translations { get; set; }
+    }
+}
