+++ ETIM-API-Saver/Common/DTO/Result/ResultOfRequestAccessToken.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Common.DTO.Result
+{
+    public class ResultOfRequestAccessToken
+    {
+        public string Access_token { get; set; }
+        public int Expires_in { get; set; }
+        public string Token_type { get; set; }
+        public string Scope { get; set; }
+    }
+}
