+++ ETIM-API-Saver/Common/DTO/Value.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Common.DTO
+{
+    public class Value
+    {
+        public string Code { get; set; }
+        public bool Deprecated { get; set; }
+        public string Description { get; set; }
+    }
+}
