+++ ETIM-API-Saver/ETIM-API-Saver.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33717.318
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ETIM-API-Saver", "ETIM-API-Saver\ETIM-API-Saver.csproj", "{CB92A9B3-ABF5-4AB0-9015-308DA3311854}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ETIM-API-Saver", "ETIM-API-Saver\ETIM-API-Saver.csproj", "{CB92A9B3-ABF5-4AB0-9015-308DA3311854}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "Common\Common.csproj", "{00B82F7A-1E83-40AF-A9C2-7DC3D657D9A6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{CB92A9B3-ABF5-4AB0-9015-308DA3311854}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CB92A9B3-ABF5-4AB0-9015-308DA3311854}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CB92A9B3-ABF5-4AB0-9015-308DA3311854}.Release|Any CPU.Build.0 = Release|Any CPU
+		{00B82F7A-1E83-40AF-A9C2-7DC3D657D9A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{00B82F7A-1E83-40AF-A9C2-7DC3D657D9A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{00B82F7A-1E83-40AF-A9C2-7DC3D657D9A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{00B82F7A-1E83-40AF-A9C2-7DC3D657D9A6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
